+++ src/pocketmine/network/rcon/RCON.php
@@ -94,7 +94,7 @@ class RCON{
 		$this->server->getTickSleeper()->addNotifier($notifier, function() : void{
 			$this->check();
 		});
-		$this->instance = new RCONInstance($this->socket, $password, (int) max(1, $maxClients), $this->server->getLogger(), $this->ipcThreadSocket, $notifier);
+		$this->instance = new RCONInstance($this->socket, $password, max(1, $maxClients), $this->server->getLogger(), $this->ipcThreadSocket, $notifier);
 
 		socket_getsockname($this->socket, $addr, $port);
 		$this->server->getLogger()->info("RCON running on $addr:$port");
